fix: Cambie el tipo de los id a Integer en el service de Notificacion

+++ src/main/java/com/git/gestion_turnos/service/INotificacion.java
@@ -11,11 +11,11 @@ public interface INotificacion {
 
     NotificacionDTO crearRecordatorio24h(Persona persona, Turno turno);
 
-    void marcarComoEnviada(Long notificacionId);
+    void marcarComoEnviada(Integer notificacionId);
 
-    void marcarComoRespondida(Long notificacionId);
+    void marcarComoRespondida(Integer notificacionId);
 
-    List<NotificacionDTO> findByPersona(Long personaId);
+    List<NotificacionDTO> findByPersona(Integer personaId);
 
-    NotificacionDTO findById(Long id);
+    NotificacionDTO findById(Integer id);
 }
